Use ssize_t for the return type of read/write instead of int.
Ported from upstream commit by crigler.

+++ dtattach.c
@@ -271,7 +271,7 @@ attach_main()
 		/* Pty activity */
 		if (n > 0 && FD_ISSET(s, &readfds))
 		{
-			int len = read(s, buf, sizeof(buf));
+			ssize_t len = read(s, buf, sizeof(buf));
 
 			if (len == 0)
 			{
